# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Compte API",
        "description": "API Documentation pour la gestion des comptes utilisateurs",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/signalements": {
            "get": {
                "tags": [
                    "Signalements"
                ],
                "summary": "Obtenir tous les signalements",
                "operationId": "91aae7b6c7f20bb9a25d8f826f47eda3",
                "responses": {
                    "200": {
                        "description": "Opération réussie",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Signalement"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/signalementStore{id}": {
            "post": {
                "tags": [
                    "Signalements"
                ],
                "summary": "Soumettre un signalement pour une annonce",
                "operationId": "9e33ed33546eee2982271a2e5e09d6f7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de l'annonce signalée",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "description": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Signalement soumis avec succès"
                    },
                    "404": {
                        "description": "Annonce non trouvée"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/signalementProprietaire": {
            "get": {
                "tags": [
                    "Signalements"
                ],
                "summary": "Obtenir les signalements liés aux annonces acceptées par l'utilisateur",
                "operationId": "00437fd38be44ff1dde7f73c8b46d498",
                "responses": {
                    "200": {
                        "description": "Opération réussie",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Signalement"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Aucun signalement trouvé"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/signalementDestroy{id}": {
            "delete": {
                "tags": [
                    "Signalements"
                ],
                "summary": "Supprimer un signalement",
                "operationId": "d6dd3b77cc9cfd1596223bd2aa4f76eb",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID du signalement à supprimer",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Signalement supprimé avec succès"
                    },
                    "404": {
                        "description": "Signalement non trouvé"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}